db/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.29
openai>=1.0.0

# Vector store and embeddings
faiss-cpu>=1.7.4
numpy>=1.21.0

# Utilities
tqdm>=4.64.0
pathlib2>=2.3.7

# Optional: If you need GPU support for FAISS
# faiss-gpu>=1.7.4  # Uncomment if using GPU

# Development dependencies (optional)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=4.0.0